Monster commit because I'm very bad at vc
PIXI is integrated and pretty much working.

No, this code doesn't compile. It's become
clear that to eventually transition to a multiplayer
server-client architecture, some pretty significant
refactoring is going to have to happen. I'd rather
do that now than later, even if it means leaving
things in a very gross ugly broken state for now.